extract focusgroup utility functions
